Update WinPreManage_browsers_history.ps1
Feature: Add User Interaction to Browser History Backup Class

This update introduces a new feature to the `BrowserHistoryBackup` PowerShell class, part of the WinPreManage suite. The class now includes a method `Invoke-UserInteraction`, which prompts the user for input directly through the console. This enhancement allows for a more interactive and user-friendly approach to specifying backup options, including the selection of browsers for history backup and the specification of backup and log file paths.

Details:
- Added `Invoke-UserInteraction` method for console-based user prompts.
- Users can now specify backup paths and choose which browsers to back up directly.
- The class initializes logging based on user-specified paths, ensuring that all activity and errors are logged.
- Enhanced feedback by logging to files and providing console output, improving the utility's usability and monitoring capabilities.

Usage:
- To engage the new interactive mode, instantiate the `BrowserHistoryBackup` class and call the `Invoke-UserInteraction` method. This method can be commented out if not needed, allowing for flexibility in automated scripts.

Note:
- This update is designed to improve user experience and flexibility, making the browser history backup process more accessible and customizable. It's especially useful in environments where dynamic input from the user or administrator is preferred.

Author: Tristan McGowan
Date: March 10, 2024
Version: 0.1.1